feat: Add new rule 'UseUTF8EncodingForHelpFile'
This commit adds a new rule 'UseUTF8EncodingForHelpFile' to the PSScriptAnalyzerSettings. This rule checks if UTF-8 encoding is used for help files.

feat: Update 'IgnoreAssignmentOperatorInsideHashTable' rule

This commit updates the 'IgnoreAssignmentOperatorInsideHashTable' rule in the PSScriptAnalyzerSettings. The rule is now set to $false, meaning it will no longer ignore assignment operators inside hash tables.